Move CLI tests to dashboard.e2e.test.js
