--- Parse-Dashboard/tests/.eslintrc.json
@@ -1,8 +1,0 @@
-{
-  "plugins": [
-    "jest"
-  ],
-  "env": {
-    "jest/globals": true
-  }
-}
