+++ package.json
@@ -143,8 +143,7 @@
   "main": "Parse-Dashboard/app.js",
   "jest": {
     "roots": [
-      "src/lib",
-      "Parse-Dashboard"
+      "src/lib"
     ],
     "transform": {
       ".*": "<rootDir>/testing/preprocessor.js"
